+++ Source/Provider Libraries/Subscription Handling/SubscriptionHandler/Interfaces.cs
@@ -7,7 +7,7 @@ namespace Hidwizards.IOWrapper.Libraries.SubscriptionHandlers
     {
         void Subscribe(InputSubscriptionRequest subReq);
         void Unsubscribe(InputSubscriptionRequest subReq);
-        bool FireCallbacks(BindingDescriptor bindingDescriptor, int value);
+        bool FireCallbacks(BindingDescriptor bindingDescriptor, short value);
     }
 
     public interface ISubscriptionInfo
